pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.72.0",
    "click>=8.2.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.19.0",
    "openai>=1.91.0",
    "python-dotenv>=1.1.1",
]
